% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ptLasso.R
\name{coef.cv.ptLasso}
\alias{coef.cv.ptLasso}
\title{Get the coefficients from a fitted cv.ptLasso model.}
\usage{
\method{coef}{cv.ptLasso}(fit, model = c("all", "individual", "overall", "pretrain"), alpha = NULL)
}
\arguments{
\item{fit}{fitted \code{"cv.ptLasso"} object.}

\item{model}{string indicating which coefficients to retrieve. Must be one of "all", "individual", "overall" or "pretrain".}

\item{alpha}{value betweein 0 and 1, indicating which alpha to use. If \code{NULL}, return coefficients from all models.  Only impacts the results for model = "all" or model = "pretrain".}
}
\description{
Get the coefficients from a fitted cv.ptLasso model.
}
\seealso{
\code{cv.ptLasso}, \code{ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
