% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ptLasso.R
\name{coef.ptLasso}
\alias{coef.ptLasso}
\title{Get the coefficients from a fitted ptLasso model.}
\usage{
\method{coef}{ptLasso}(fit, model = c("all", "individual", "overall", "pretrain"))
}
\arguments{
\item{fit}{fitted \code{"ptLasso"} object.}

\item{model}{string indicating which coefficients to retrieve. Must be one of "all", "individual", "overall" or "pretrain".}
}
\description{
Get the coefficients from a fitted ptLasso model.
}
\seealso{
\code{ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
