% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ptLasso.R
\name{cv.ptLasso}
\alias{cv.ptLasso}
\title{Cross-validation for ptLasso}
\usage{
cv.ptLasso(
  x,
  y,
  w = rep(1, length(y)),
  alphalist = seq(0, 1, length = 11),
  groups = NULL,
  family = c("gaussian", "multinomial", "binomial", "cox"),
  useCase = c("inputGroups", "targetGroups"),
  type.measure = c("default", "mse", "mae", "auc", "deviance", "class", "C"),
  nfolds = 10,
  foldid = NULL,
  keep = FALSE,
  verbose = FALSE,
  fitall = NULL,
  fitind = NULL,
  trace.it = FALSE,
  weights = NULL,
  overall.lambda = "lambda.1se",
  s = "lambda.min",
  parallel = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{x} matrix as in \code{ptLasso}.}

\item{y}{\code{y} matrix as in \code{ptLasso}.}

\item{groups}{\code{groups} vector as in \code{ptLasso}}

\item{family}{Response type as in \code{ptLasso}.}

\item{type.measure}{Measure computed in cv.glmnet, as in \code{ptLasso}.}

\item{verbose}{If \code{verbose=1}, print a statement showing which model is currently being fit.}

\item{...}{Other arguments that can be passed to \code{ptLasso}.}

\item{alpha}{A vector of values of the pretraining hyperparameter alpha. Defaults to \code{seq(0, 1, length.out=11)}.}
}
\value{
An object of class \code{"cv.ptLasso"}, which is a list with the
ingredients of the cross-validation fit.
\item{glmfit}{A fitted \code{ptLasso} object for the full data.}
\item{theta}{Value of \code{theta} used in model fitting.}
\item{lambda}{The values of \code{lambda} used in the fits.}
\item{nzero}{If the groups overlap, the number of non-zero coefficients
  in the model \code{glmfit} for the expanded feature space, at each value of
  \code{lambda}. Otherwise, the number of non-zero coefficients in the model
  \code{glmfit} for the original feature space.}
\item{orignzero}{If the groups are overlapping, this is the number of
  non-zero coefficients in the model \code{glmfit} for the original feature
  space, at each \code{lambda}. If groups are not overlapping, it is
  \code{NULL}.}
\item{fit.preval}{If \code{keep=TRUE}, this is the array of prevalidated
  fits.}
\item{cvm}{The mean cross-validated error: a vector of length
  \code{length(lambda)}.}
\item{cvse}{Estimate of standard error of \code{cvm}.}
\item{cvlo}{Lower curve = \code{cvm - cvsd}.}
\item{cvup}{Upper curve = \code{cvm + cvsd}.}
\item{lambda.min}{The value of \code{lambda} that gives minimum
  \code{cvm}.}
\item{lambda.1se}{The largest value of \code{lambda} such that the CV
  error is within one standard error of the minimum.}
\item{foldid}{If \code{keep=TRUE}, the fold assignments used.}
\item{name}{Name of error measurement used for CV.}
\item{call}{The call that produced this object.}
}
\description{
Cross-validation for \code{ptLasso}.
}
\details{
This function runs \code{ptLasso} once for each choice of alpha, and returns the cross-validated performance computed by cv.glmnet within each call to ptLasso.
}
\examples{
1+1

}
\seealso{
\code{\link{ptLasso}} and \code{\link{plot.cv.ptLasso}}.
}
