% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{gaussian.example.data}
\alias{gaussian.example.data}
\title{Simulate input grouped data (gaussian outcome) for testing with ptLasso.}
\usage{
gaussian.example.data(
  k = 5,
  class.sizes = rep(200, k),
  n = sum(class.sizes),
  scommon = 10,
  sindiv = rep(10, k),
  p = 2 * (sum(sindiv) + scommon),
  beta.common = 3 * (1:k),
  beta.indiv = rep(3, k),
  intercepts = rep(0, k),
  sigma = 20
)
}
\arguments{
\item{k}{Default: 5.}

\item{class.sizes}{Default: rep(200, k).}

\item{n}{Default: sum(class.sizes).}

\item{scommon}{Default: 10.}

\item{sindiv}{Default: rep(10, k).}

\item{p}{Default: 2*(sum(sindiv) + scommon).}

\item{beta.common}{Default: 3*(1:k).}

\item{beta.indiv}{Default: rep(3, k).}

\item{intercepts}{Default: rep(0, k).}

\item{sigma}{Default: 20.}
}
\value{
A list for data with 5 groups and a gaussian outcome, n=1000 and p=120:
\item{x}{Simulated features, size n x p.}
\item{y}{Outcomes y, length n.}
\item{groups}{Vector of length n, indicating which observations belong to which group.}
\item{snr}{Gaussian outcome only: signal to noise ratio.}
\item{mu}{Gaussian outcome only: the value of y before noise is added.}
}
\description{
No required arguments; used primarily for documentation. Simply calls \code{makedata} with a reasonable set of features.
}
\examples{
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group

}
\seealso{
\code{cv.ptLasso}, \code{ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
