% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ptLasso.R
\name{get.individual.support}
\alias{get.individual.support}
\title{Get the support for individual models}
\usage{
get.individual.support(
  fit,
  s = "lambda.min",
  gamma = "gamma.min",
  commonOnly = FALSE,
  groups = 1:length(fit$fitind)
)
}
\arguments{
\item{fit}{fitted \code{"ptLasso"} or \code{"cv.ptLasso"} object.}

\item{s}{the choice of lambda to use. May be "lambda.min", "lambda.1se" or a numeric value. Default is "lambda.min".}

\item{gamma}{for use only when 'relax = TRUE' was specified during training. The choice of 'gamma' to use. May be "gamma.min" or "gamma.1se". Default is "gamma.min".}

\item{commonOnly}{whether to return the features that are chosen by more than half of the group- or response-specific models (TRUE) or the features that are chosen by any of the group-specific models (FALSE). Default is FALSE.}

\item{groups}{which groups or responses to include when computing the support. Default is to include all groups/responses.}
}
\value{
This returns a vector containing the indices of nonzero coefficients (excluding the intercept).
}
\description{
Get the indices of nonzero coefficients from the individual models in a fitted ptLasso or cv.ptLasso object, excluding the intercept.
}
\examples{
# Train data
set.seed(1234)
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group;

fit = ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")

get.individual.support(fit) 

# only return features common to all groups 
get.individual.support(fit, commonOnly = TRUE) 

# group 1 only
get.individual.support(fit, groups = 1) 

cvfit = cv.ptLasso(x, y, groups = groups, alphalist = c(0, .5, 1), 
                   family = "gaussian", type.measure = "mse")

get.individual.support(cvfit)

# group 1 only
get.individual.support(cvfit, groups = 1) 
 
}
\seealso{
\code{ptLasso}, \code{cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
