% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.ptLasso.R
\name{plot.predict.cv.ptLasso}
\alias{plot.predict.cv.ptLasso}
\title{Plot the performance curve produced by predict.cv.ptLasso using a validation set, as a function of the \code{alpha} values used.}
\usage{
\method{plot}{predict.cv.ptLasso}(x, plot.alphahat = TRUE, ...)
}
\arguments{
\item{x}{\code{"predict.cv.ptLasso"} object.}

\item{plot.alphahat}{If \code{TRUE}, show a dashed vertical line indicating the single value of alpha that maximized overall cross-validated performance.}

\item{\dots}{Other graphical parameters to plot}
}
\description{
A plot is produced, and nothing is returned.
}
\examples{
3+3


}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
