% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ptLasso.R
\name{predict.cv.ptLasso}
\alias{predict.cv.ptLasso}
\title{Predict using a cv.ptLasso object.}
\usage{
\method{predict}{cv.ptLasso}(
  cvfit,
  xtest,
  groupstest = NULL,
  ytest = NULL,
  alpha = NULL,
  alphatype = c("fixed", "varying"),
  type = c("link", "response", "class"),
  s = "lambda.min"
)
}
\arguments{
\item{cvfit}{Fitted \code{"cv.ptLasso"} object.}

\item{xtest}{Input matrix, matching the form used by \code{"cv.ptLasso"} for model training.}

\item{groupstest}{A vector indicating to which group each observation belongs. Coding should match that used for model training. Will be NULL for target grouped data.}

\item{ytest}{Response variable. Optional. If included, \code{"predict"} will compute performance measures for xtest using code{"type.measure"} from the cvfit object.}

\item{alpha}{The chosen alpha to use for prediction. If NULL, this will rely on the choice of "alphatype".}

\item{alphatype}{Choice of '"fixed"' or '"varying"'. If '"fixed"', use the alpha that achieved best cross-validated performance. If '"varying"', each group uses the alpha that optimized the group-specific cross-validated performance.}

\item{type}{Type of prediction required. Type '"link"' gives the linear predictors for '"binomial", '"multinomial"' or '"cox"' models; for '"gaussian"' models it gives the fitted values. Type '"response"' gives the fitted probabilities for '"binomial"' or '"multinomial"', and the fitted relative-risk for '"cox"'; for '"gaussian"' type '"response"' is equivalent to type '"link"'. Note that for '"binomial"' models, results are returned only for the class corresponding to the second level of the factor response. Type '"class"' applies only to '"binomial"' or '"multinomial"' models, and produces the class label corresponding to the maximum probability.}

\item{s}{Value of the penalty parameter 'lambda' at which predictions are required. Will use the same lambda for all models; can be a numeric value, '"lambda.min"' or '"lambda.1se"'. Default is '"lambda.min"'.}
}
\value{
A ...
}
\description{
Return predictions and performance measures for a test set.
}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
