% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.cv.ptLasso}
\alias{print.cv.ptLasso}
\title{Print the cv.ptLasso object.}
\usage{
\method{print}{cv.ptLasso}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{"cv.ptLasso"} or \code{"predict.cv.ptLasso"} object.}

\item{\dots}{Other graphical parameters to plot}
}
\description{
Print the cv.ptLasso object.
}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
