% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.predict.ptLasso}
\alias{print.predict.ptLasso}
\title{Print the predict.ptLasso object.}
\usage{
\method{print}{predict.ptLasso}(x, ...)
}
\arguments{
\item{x}{output of predict called with a ptLasso object.}

\item{\dots}{other arguments to pass to the print function.}
}
\description{
Print the predict.ptLasso object.
}
\examples{
# Train data
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group;

# Test data
outtest = gaussian.example.data()
xtest=outtest$x; ytest=outtest$y; groupstest=outtest$groups

fit = ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")
pred = predict(fit, xtest, groupstest, ytest=ytest, s="lambda.min")
print(pred)

# If ytest is not supplied, just prints the pretrained predictions.
pred = predict(fit, xtest, groupstest, s="lambda.min")
print(pred)

}
\seealso{
\code{ptLasso} and \code{predict.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
