% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ptLasso.R
\name{coef.cv.ptLasso}
\alias{coef.cv.ptLasso}
\title{Get the coefficients from a fitted cv.ptLasso model.}
\usage{
\method{coef}{cv.ptLasso}(
  object,
  model = c("all", "individual", "overall", "pretrain"),
  alpha = NULL,
  ...
)
}
\arguments{
\item{object}{fitted \code{"cv.ptLasso"} object.}

\item{model}{string indicating which coefficients to retrieve. Must be one of "all", "individual", "overall" or "pretrain".}

\item{alpha}{value between 0 and 1, indicating which alpha to use. If \code{NULL}, return coefficients from all models.  Only impacts the results for model = "all" or model = "pretrain".}

\item{\dots}{other arguments to be passed to the \code{"coef"} function. May be e.g. \code{s = "lambda.min"}.}
}
\description{
Get the coefficients from a fitted cv.ptLasso model.
}
\examples{
set.seed(1234)
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group;

cvfit = cv.ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")
# Get all model coefficients.
names(coef(cvfit))

coef(cvfit, model = "overall") # Overall model only
length(coef(cvfit, model = "individual")) # List of coefficients for each group model
length(coef(cvfit, model = "pretrain", alpha = .5)) # List of coefficients for each group model

}
\seealso{
\code{cv.ptLasso}, \code{ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
