% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ptLasso.R
\name{coef.ptLasso}
\alias{coef.ptLasso}
\title{Plot the cross-validation (or validation) curve produced by cv.ptLasso (or predict.cv.ptLasso), as a function of the \code{alpha} values used.}
\usage{
\method{coef}{ptLasso}(fit, model = c("all", "individual", "overall", "pretrain"))
}
\arguments{
\item{fit}{fitted \code{"ptLasso"} object.}

\item{model}{string indicating which coefficients to retrieve. Must be one of "all", "individual", "overall" or "pretrain".}
}
\description{
A plot is produced, and nothing is returned.
}
\seealso{
\code{ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
