% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ptLasso.R
\name{cv.ptLasso}
\alias{cv.ptLasso}
\title{Cross-validation for ptLasso}
\usage{
cv.ptLasso(
  x,
  y,
  w = rep(1, length(y)),
  alphalist = seq(0, 1, length = 11),
  groups = NULL,
  family = c("gaussian", "multinomial", "binomial", "cox"),
  use.case = c("inputGroups", "targetGroups"),
  type.measure = c("default", "mse", "mae", "auc", "deviance", "class", "C"),
  nfolds = 10,
  foldid = NULL,
  keep = FALSE,
  verbose = FALSE,
  fitall = NULL,
  fitind = NULL,
  trace.it = FALSE,
  weights = NULL,
  overall.lambda = "lambda.1se",
  s = "lambda.min",
  parallel = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{x} matrix as in \code{ptLasso}.}

\item{y}{\code{y} matrix as in \code{ptLasso}.}

\item{groups}{\code{groups} vector as in \code{ptLasso}}

\item{family}{Response type as in \code{ptLasso}.}

\item{type.measure}{Measure computed in \code{cv.glmnet}, as in \code{ptLasso}.}

\item{verbose}{If \code{verbose=1}, print a statement showing which model is currently being fit.}

\item{...}{Other arguments that can be passed to \code{ptLasso}.}

\item{alpha}{A vector of values of the pretraining hyperparameter alpha. Defaults to \code{seq(0, 1, length.out=11)}.}
}
\value{
An object of class \code{"cv.ptLasso"}, which is a list with the ingredients of the cross-validation fit.
\item{fitall}{A fitted \code{cv.glmnet} object trained using the full data.}
\item{fitpre}{A list of fitted (pretrained) \code{cv.glmnet} objects, one trained with each group.}
\item{fitind}{A list of fitted \code{cv.glmnet} objects, one trained with each group.}
\item{alphahat}{Value of \code{alpha} that optimizes CV performance on all data.}
\item{varying.alphahat}{Vector of values of \code{alpha}, the kth of which optimizes performance for group k.}
\item{alphalist}{Vector of all alphas that were compared.}
\item{errall}{CV performance for the overall model.}
\item{errpre}{CV performance for the pretrained models (one for each \code{alpha} tried).}
\item{errind}{CV performance for the individual model.}
\item{useCase}{Whether the objective is "inputGroups" or "targetGroups".}
\item{type.measure}{Measure computed in \code{cv.glmnet}.}
\item{family}{Response type.}
\item{fit}{List of \code{ptLasso} objects, one for each \code{alpha} tried.}
}
\description{
Cross-validation for \code{ptLasso}.
}
\details{
This function runs \code{ptLasso} once for each choice of alpha, and returns the cross-validated performance computed by cv.glmnet within each call to ptLasso.
}
\examples{
1+1

}
\seealso{
\code{\link{ptLasso}} and \code{\link{plot.cv.ptLasso}}.
}
