% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ptLasso.R
\name{get.overall.support}
\alias{get.overall.support}
\title{Get the support for the overall model}
\usage{
get.overall.support(fit, s = "lambda.min", gamma = "gamma.min")
}
\arguments{
\item{fit}{fitted \code{"ptLasso"} or \code{"cv.ptLasso"} object.}

\item{s}{the choice of lambda to use. May be "lambda.min", "lambda.1se" or a numeric value. Default is "lambda.min".}

\item{gamma}{for use only when 'relax = TRUE' was specified during training. The choice of 'gamma' to use. May be "gamma.min" or "gamma.1se". Default is "gamma.min".}
}
\value{
This returns a vector containing the indices of nonzero coefficients (excluding the intercept).
}
\description{
Get the indices of nonzero coefficients from the overall model in a fitted ptLasso or cv.ptLasso object, excluding the intercept.
}
\examples{
# Train data
set.seed(1234)
out = gaussian.example.data(k=2, class.sizes = c(50, 50))
x = out$x; y=out$y; groups = out$group;

fit = ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")

get.overall.support(fit, s="lambda.min") 
get.overall.support(fit, s="lambda.1se") 

cvfit = cv.ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")

get.overall.support(cvfit, s="lambda.min") 
get.overall.support(cvfit, s="lambda.1se") 
 
}
\seealso{
\code{ptLasso}, \code{cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
