% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{makedata.targetgroups}
\alias{makedata.targetgroups}
\title{Simulate target grouped data for testing with ptLasso.}
\usage{
makedata.targetgroups(
  n,
  p,
  scommon,
  sindiv,
  class.sizes,
  shift.common,
  shift.indiv
)
}
\arguments{
\item{n}{Total number of observations to simulate.}

\item{p}{Total number of features to simulate.}

\item{scommon}{Number of features shared by all groups.}

\item{sindiv}{Vector of length k. The i^th entry indicates the number of features specific to group i.}

\item{class.sizes}{Vector of length k. The i^th entry indicates the number of observations in group i.}

\item{shift.common}{A list of length k (one for each group). Each entry of this list should be a vector of length scommon, containing the shifts for the scommon features. The i^th entry of this list will be added to the first scommon columns of x for observations in group i.}

\item{shift.indiv}{The shifts for the individual features, in the same form as shift.common.}
}
\value{
A list:
\item{x}{Simulated features, size n x p.}
\item{y}{Outcomes y, length n.}
}
\description{
Simulate target grouped data for testing with ptLasso.
}
\examples{

k = 5
class.sizes = rep(50, k)
n = sum(class.sizes)
scommon = 3
sindiv = rep(3, k)
p = 3*(sum(sindiv) + scommon)
shift.common  = lapply(seq(-.1, .1, length.out = k), function(i) rep(i, scommon))
shift.indiv = lapply(1:k, function(i) -shift.common[[i]])

out = makedata.targetgroups(n=n, p=p, scommon=scommon,
                            sindiv=sindiv, class.sizes=class.sizes,
                            shift.common=shift.common, shift.indiv=shift.indiv)
x = out$x; y=out$y


}
\seealso{
\code{cv.ptLasso}, \code{ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
