% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.ptLasso.R
\name{plot.ptLasso}
\alias{plot.ptLasso}
\title{Plot the models trained by a ptLasso object}
\usage{
\method{plot}{ptLasso}(fit)
}
\arguments{
\item{fit}{Fitted \code{"ptLasso"} object.}
}
\description{
A plot is produced, and nothing is returned.
}
\examples{
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group

fit = ptLasso(x, y, groups = groups, alpha = 0.5, family = "gaussian", type.measure = "mse")
plot(fit) 

}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer: Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
