% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ptLasso.R
\name{predict.cv.ptLasso}
\alias{predict.cv.ptLasso}
\title{Predict using a cv.ptLasso object.}
\usage{
\method{predict}{cv.ptLasso}(
  cvfit,
  xtest,
  groupstest = NULL,
  ytest = NULL,
  alpha = NULL,
  alphatype = c("fixed", "varying"),
  type = c("link", "response", "class"),
  s = "lambda.min",
  return.link = FALSE
)
}
\arguments{
\item{cvfit}{Fitted \code{"cv.ptLasso"} object.}

\item{xtest}{Input matrix, matching the form used by \code{"cv.ptLasso"} for model training.}

\item{groupstest}{A vector indicating to which group each observation belongs. Coding should match that used for model training. Will be NULL for target grouped data.}

\item{ytest}{Response variable. Optional. If included, \code{"predict"} will compute performance measures for xtest using code{"type.measure"} from the cvfit object.}

\item{alpha}{The chosen alpha to use for prediction. May be a vector containing one value of alpha for each group. If NULL, this will rely on the choice of "alphatype".}

\item{alphatype}{Choice of '"fixed"' or '"varying"'. If '"fixed"', use the alpha that achieved best cross-validated performance. If '"varying"', each group uses the alpha that optimized the group-specific cross-validated performance.}

\item{type}{Type of prediction required. Type '"link"' gives the linear predictors for '"binomial", '"multinomial"' or '"cox"' models; for '"gaussian"' models it gives the fitted values. Type '"response"' gives the fitted probabilities for '"binomial"' or '"multinomial"', and the fitted relative-risk for '"cox"'; for '"gaussian"' type '"response"' is equivalent to type '"link"'. Note that for '"binomial"' models, results are returned only for the class corresponding to the second level of the factor response. Type '"class"' applies only to '"binomial"' or '"multinomial"' models, and produces the class label corresponding to the maximum probability.}

\item{s}{Value of the penalty parameter 'lambda' at which predictions are required. Will use the same lambda for all models; can be a numeric value, '"lambda.min"' or '"lambda.1se"'. Default is '"lambda.min"'.}

\item{return.link}{If \code{TRUE}, will additionally return the linear link for the overall, pretrained and individual models: \code{linkoverall}, \code{linkpre} and \code{linkind}.}
}
\value{
A list containing the requested predictions. If \code{ytest} is included, will also return error measures.
\item{call}{The call that produced this object.}
\item{alpha}{The value(s) of alpha used to generate predictions.}
\item{yhatoverall}{Predictions from the overall model.}
\item{yhatind}{Predictions from the individual models.}
\item{yhatpre}{Predictions from the pretrained models.}
\item{supoverall}{Indices of the features selected by the overall model.}
\item{supind}{Union of the indices of the features selected by the individual models.}
\item{suppre.common}{Features selected in the first stage of pretraining.}
\item{suppre.individual}{Union of the indices of the features selected by the pretrained models, without the features selected in the first stage.}
\item{type.measure}{If \code{ytest} is supplied, the performance measure computed.}
\item{erroverall}{If \code{ytest} is supplied, performance for the overall model. This is a named vector containing performance for (1) the entire dataset, (2) the average performance across groups, (3) the average performance across groups weighted by group size and (4) group-specific performance.}
\item{errind}{If \code{ytest} is supplied, performance for the overall model. As described in \code{erroverall}.}
\item{errpre}{If \code{ytest} is supplied, performance for the overall model. As described in \code{erroverall}.}
\item{linkoverall}{If \code{return.link} is TRUE, return the linear link from the overall model.}
\item{linkind}{If \code{return.link} is TRUE, return the linear link from the individual models.}
\item{linkpre}{If \code{return.link} is TRUE, return the linear link from the pretrained models.}
}
\description{
Return predictions and performance measures for a test set.
}
\examples{
#### Gaussian example
set.seed(1234)
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group;
outtest = gaussian.example.data()
xtest=outtest$x; ytest=outtest$y; groupstest=outtest$groups;

# Model fitting
# By default, use the single value of alpha that had the best CV performance on the entire dataset:
cvfit = cv.ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")
pred = predict(cvfit, xtest, groupstest, ytest=ytest, s="lambda.min")
pred

# For each group, use the value of alpha that had the best CV performance for that group:
pred = predict(cvfit, xtest, groupstest, ytest=ytest, s="lambda.min", alphatype = "varying")
pred

# Specify a single value of alpha and use lambda.1se.
pred = predict(cvfit, xtest, groupstest, ytest=ytest, s="lambda.1se",
               alphatype = "varying", alpha = .3)
pred

# Specify a vector of choices for alpha: 
pred = predict(cvfit, xtest, groupstest, ytest=ytest, s="lambda.min",
               alphatype = "varying", alpha = c(.1, .2, .3, .4, .5))
pred

}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
