% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ptLasso.R
\name{predict.ptLasso}
\alias{predict.ptLasso}
\title{Predict using a ptLasso object.}
\usage{
\method{predict}{ptLasso}(
  fit,
  xtest,
  groupstest = NULL,
  ytest = NULL,
  offset = NULL,
  type = c("link", "response", "class"),
  s = "lambda.min"
)
}
\arguments{
\item{xtest}{Input matrix, matching the form used by \code{"ptLasso"} for model training.}

\item{groupstest}{A vector indicating to which group each observation belongs. Coding should match that used for model training. Will be NULL for target grouped data.}

\item{type}{Type of prediction required. Type '"link"' gives the linear predictors for '"binomial", '"multinomial"' or '"cox"' models; for '"gaussian"' models it gives the fitted values. Type '"response"' gives the fitted probabilities for '"binomial"' or '"multinomial"', and the fitted relative-risk for '"cox"'; for '"gaussian"' type '"response"' is equivalent to type '"link"'. Note that for '"binomial"' models, results are returned only for the class corresponding to the second level of the factor response. Type '"class"' applies only to '"binomial"' or '"multinomial"' models, and produces the class label corresponding to the maximum probability.}

\item{s}{Value of the penalty parameter 'lambda' at which predictions are required. Will use the same lambda for all models; can be a numeric value, '"lambda.min"' or '"lambda.1se"'. Default is '"lambda.min"'.}

\item{cvfit}{Fitted \code{"ptLasso"} object.}
}
\value{
A ..
}
\description{
Return predictions and performance measures for a test set.
}
