% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.cv.ptLasso}
\alias{print.cv.ptLasso}
\title{Print the cv.ptLasso object.}
\usage{
\method{print}{cv.ptLasso}(x, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{fitted \code{"cv.ptLasso"} object.}

\item{digits}{number of digits to display.}
}
\description{
Print the cv.ptLasso object.
}
\examples{
out = gaussian.example.data()
x = out$x; y=out$y; groups = out$group;

cvfit = cv.ptLasso(x, y, groups = groups, family = "gaussian", type.measure = "mse")
print(cvfit)

}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
