% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.predict.cv.ptLasso}
\alias{print.predict.cv.ptLasso}
\title{Print the predict.ptLasso object.}
\usage{
\method{print}{predict.cv.ptLasso}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{output of predict called with a ptLasso object.}

\item{digits}{number of digits to display.}
}
\description{
Print the predict.ptLasso object.
}
\seealso{
\code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
