% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.ptLasso}
\alias{print.ptLasso}
\title{Print the ptLasso object.}
\usage{
\method{print}{ptLasso}(x)
}
\arguments{
\item{x}{fitted \code{"ptLasso"} object.}
}
\description{
Print the ptLasso object.
}
\seealso{
\code{ptLasso}, \code{cv.ptLasso} and \code{predict.cv.ptLasso}.
}
\author{
Erin Craig and Rob Tibshirani\cr Maintainer:
Erin Craig <erincr@stanford.edu>
}
\keyword{classification}
\keyword{models}
\keyword{regression}
